requirements.txt
# Core Django
Django==4.2
python-decouple==3.8
Pillow==10.2.0
jazzmin

# Authentication & Security
django-allauth==0.61.1
django-crispy-forms==2.1
python-jose==3.3.0

# Database
psycopg2-binary==2.9.9

# Cache & Performance
#django-redis==5.4.0
#redis==5.0.1
django-compressor==4.4
sorl-thumbnail==12.10.0

# Payment & Crypto
mnemonic==0.20
#pysha3==1.0.2
#web3==6.15.1

# Utils
python-dotenv==1.0.1
#django-environ==0.11.2
#django-storages==1.14.2

# Development & Debug
#django-debug-toolbar==4.3.0
#black==24.2.0
#flake8==7.0.0

# Testing
#pytest==8.0.2
#pytest-django==4.8.0
#coverage==7.4.1

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0